requirements.txt
autopep8==2.0.2
imageio==2.27.0
lazy_loader==0.2
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pycodestyle==2.10.0
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.10.1
tifffile==2023.4.12
tomli==2.0.1
